# backend/docs/swagger.yaml
definitions:
  handlers.MattermostRequest:
    properties:
      message:
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      color:
        type: string
      login:
        type: string
      password:
        type: string
    type: object
  handlers.UserResponse:
    properties:
      color:
        type: string
      login:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрирует пользователя
      tags:
      - users
  /api/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получает всех пользователей
      tags:
      - users
  /send-to-mattermost:
    post:
      consumes:
      - application/json
      parameters:
      - description: Сообщение
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.MattermostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Шлёт сообщение в Mattermost
      tags:
      - mattermost
swagger: "2.0"